+++ vite.config.js
@@ -3,6 +3,7 @@ import { resolve } from "path";
 import { defineConfig } from "vite";
 import dts from "vite-plugin-dts";
 import { viteStaticCopy } from "vite-plugin-static-copy";
+import scss from 'rollup-plugin-scss';
 
 export default defineConfig({
   build: {
@@ -11,6 +12,8 @@ export default defineConfig({
       // Could also be a dictionary or array of multiple entry points
       entry: [
         resolve(__dirname, "src/index.ts"),
+        resolve(__dirname, "src/style.scss"),
+        resolve(__dirname, "src/baseline.scss"),
         resolve(__dirname, "src/components/badge/badge.ts"),
         resolve(__dirname, "src/components/title/title.ts"),
       ],
@@ -19,6 +22,7 @@ export default defineConfig({
       // the proper extensions will be added
       fileName: "[name]",
     },
+    cssCodeSplit: true,
     rollupOptions: {
       // make sure to externalize deps that shouldn't be bundled
       // into your library
@@ -34,21 +38,16 @@ export default defineConfig({
   },
   optimizeDeps: {},
   plugins: [
+    // viteStaticCopy({
+    //   targets: [
+    //     {
+    //       src: "LICENSE",
+    //       dest: "",
+    //     },
+    //   ],
+    // }),
     dts({
       rollupTypes: true,
     }),
-    viteStaticCopy({
-      targets: [
-        {
-          src: "src/fonts.css",
-          dest: "",
-        },
-        {
-          src: "src/assets",
-          dest: "",
-        },
-      ],
-    }),
   ],
-  // publicDir: 'assets',
 });